Refactor components so that stylesheets sit directly next to associated component for readability and ease imports. Also include index files for all component/styles folders for ease of importing in other components.
