--- src/features/converse/Converse.tsx
@@ -1,5 +1,0 @@
-import React from "react"
-
-export function Converse() {
-    return <>Converse app</>
-}
