--- src/features/diary_entry/index.tsx
@@ -1,2 +1,0 @@
-export * from './SearchDiariesPage'
-export * from './DiaryEntryPage'
