+++ src/features/diary_entry_page/index.tsx
@@ -1,0 +1,1 @@
+export * from './DiaryEntryPage'
