+++ src/features/search_diaries_page/index.tsx
@@ -1,0 +1,1 @@
+export * from './SearchDiariesPage'
