+++ src/shared/back-button/index.tsx
@@ -1,0 +1,1 @@
+export * from './BackButton'
