+++ src/shared/diary-entry-title/index.tsx
@@ -1,0 +1,1 @@
+export * from './DiaryEntryTitle'
