+++ src/shared/quote-card/QuoteCard.css
