+++ src/shared/quote-card/index.tsx
@@ -1,0 +1,1 @@
+export * from './QuoteCard'
