+++ src/shared/utils/quote_card.ts
